move FPGA DDS files into this project directory * MIDI control of single notes is working * ADSR is functional

+++ synth-wavetable.cydsn/oscillator.h
@@ -6,7 +6,7 @@ struct voice {
     uint32_t env_index;
     uint16_t env_multiplier;
     uint8_t current_env_mode;
-    int freq;
+    uint16_t freq;
     uint8_t note_index;
 };
